requirements.txt
argparse==1.2.1
httplib2==0.8
keyring==3.7
launchpadlib==1.10.2
lazr.authentication==0.1.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
oauth==1.0.1
oauthlib==0.6.1
pubsub==0.1.1
python-twitter==1.3.1
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.4.0
testresources==0.2.7
wadllib==1.3.2
wsgi-intercept==0.6.1
wsgiref==0.1.2
zope.interface==4.1.1
